feat: - 이미지 업로드 기능 개선 - 게시글 임시 저장 기능 추가 - 엘라스틱서치 연동 수정

+++ src/main/java/com/eventorback/image/service/impl/ImageServiceImpl.java
@@ -14,11 +14,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.eventorback.image.domain.dto.request.DeleteImageRequest;
+import com.eventorback.image.domain.dto.response.GetImageResponse;
 import com.eventorback.image.domain.entity.Image;
 import com.eventorback.image.exception.CreateFolderException;
 import com.eventorback.image.exception.FileExtensionException;
 import com.eventorback.image.exception.FileSaveException;
-import com.eventorback.image.exception.FileUploadException;
 import com.eventorback.image.exception.ImageNotFoundException;
 import com.eventorback.image.repository.ImageRepository;
 import com.eventorback.image.service.ImageService;
@@ -44,14 +45,17 @@ public class ImageServiceImpl implements ImageService {
 	private String uploadPath;
 
 	@Override
-	public void uploadThumbnail(MultipartFile thumbnail, String folderName, Long postId) throws FileUploadException {
+	// @Async("imageUploadExecutor")
+	public List<GetImageResponse> upload(MultipartFile file, String folderName, Long postId, boolean isThumbnail,
+		boolean isPasted) {
+
 		// 1. 날짜별 하위 폴더 생성 (YYYYMMDD 형식)
 		String today = new SimpleDateFormat("yyyyMMdd").format(new Date());
 		Path folderPath = Paths.get(uploadPath, folderName, today);
 		createDirectoryIfNotExists(folderPath);
 
 		// 2. 파일 이름 생성 (UUID 로 중복 방지)
-		String originalFilename = thumbnail.getOriginalFilename();
+		String originalFilename = file.getOriginalFilename();
 		String fileExtension = getFileExtension(originalFilename);
 		String newFileName = postId.toString() + "_" + UUID.randomUUID() + fileExtension;
 
@@ -59,55 +63,30 @@ public class ImageServiceImpl implements ImageService {
 		checkFileExtension(fileExtension);
 
 		// 4. 파일 저장
-		saveFile(folderPath, newFileName, thumbnail);
+		saveFile(folderPath, newFileName, file);
 
-		//임시 백엔드 리소스 URL
-		String url = domainUrl + folderName + "/" + today + "/" + newFileName;
+		// 리소스 URL
+		String url = domainUrl + "/" + folderName + "/" + today + "/" + newFileName;
 
 		// 기존 썸네일 존재하면 제거
-		imageRepository.deleteByPostPostIdAndIsThumbnail(postId, true);
+		if (isThumbnail) {
+			imageRepository.deleteByPostPostIdAndIsThumbnail(postId, true);
+		}
 
 		// 5. DB에 이미지 정보 저장
-		createImage(postId, originalFilename, newFileName, url, thumbnail.getSize(), true);
+		createImage(postId, originalFilename, newFileName, url, file.getSize(), isThumbnail, isPasted);
+
+		return imageRepository.getAllByPostId(postId);
 	}
 
 	@Override
-	// @Async("imageUploadExecutor")
-	public void upload(List<MultipartFile> files, String folderName, Long postId) {
-		for (MultipartFile file : files) {
-			// 1. 날짜별 하위 폴더 생성 (YYYYMMDD 형식)
-			String today = new SimpleDateFormat("yyyyMMdd").format(new Date());
-			Path folderPath = Paths.get(uploadPath, folderName, today);
-			createDirectoryIfNotExists(folderPath);
-
-			// 2. 파일 이름 생성 (UUID 로 중복 방지)
-			String originalFilename = file.getOriginalFilename();
-			String fileExtension = getFileExtension(originalFilename);
-			String newFileName = postId.toString() + "_" + UUID.randomUUID() + fileExtension;
-
-			// 3. 파일 확장자 검사
-			checkFileExtension(fileExtension);
-
-			// 4. 파일 저장
-			saveFile(folderPath, newFileName, file);
-
-			//임시 백엔드 리소스 URL
-			String url = domainUrl + folderName + "/" + today + "/" + newFileName;
-
-			// 5. DB에 이미지 정보 저장
-			createImage(postId, originalFilename, newFileName, url, file.getSize(), false);
-		}
-	}
-
-	@Override
-	public String saveFile(Path folderPath, String fileName, MultipartFile file) {
+	public void saveFile(Path folderPath, String fileName, MultipartFile file) {
 		Path filePath = folderPath.resolve(fileName);
 		try {
 			Files.write(filePath, file.getBytes());
 		} catch (IOException e) {
 			throw new FileSaveException(e.getMessage());
 		}
-		return filePath.toString();
 	}
 
 	/**
@@ -149,35 +128,55 @@ public class ImageServiceImpl implements ImageService {
 
 	@Override
 	public void createImage(Long postId, String originalName, String newName, String url, Long size,
-		Boolean isThumbnail) {
+		boolean isThumbnail, boolean isPasted) {
 		Post post = postRepository.findById(postId).orElseThrow(PostNotFoundException::new);
-		imageRepository.save(new Image(post, originalName, newName, url, size, isThumbnail));
+		imageRepository.save(new Image(post, originalName, newName, url, size, isThumbnail, isPasted));
 	}
 
 	@Override
-	public void deleteImage(Long postId, List<Long> deleteImageIds) {
-		if (deleteImageIds != null) {
-			deleteImageIds.forEach(imageId -> {
-				// 1. DB 에서 이미지 정보 조회
-				Image image = imageRepository.findById(imageId)
-					.orElseThrow(ImageNotFoundException::new);
+	public List<GetImageResponse> deleteImage(DeleteImageRequest request) {
+		request.imageIds().forEach(imageId -> {
+			// 1. DB 에서 이미지 정보 조회
+			Image image = imageRepository.findById(imageId)
+				.orElseThrow(ImageNotFoundException::new);
 
-				// 2. 실제 파일 경로 가져오기
-				Path filePath = Paths.get(uploadPath,
-					image.getUrl().replaceFirst(domainUrl, ""));
+			// 2. 실제 파일 경로 가져오기
+			Path filePath = Paths.get(uploadPath,
+				image.getUrl().replaceFirst(domainUrl, ""));
 
-				try {
-					// 3. 실제 파일 삭제
-					Files.deleteIfExists(filePath);
-				} catch (IOException e) {
-					log.error("파일 삭제 실패: {}", filePath, e);
-				}
+			try {
+				// 3. 실제 파일 삭제
+				Files.deleteIfExists(filePath);
+			} catch (IOException e) {
+				log.error("파일 삭제 실패: {}", filePath, e);
+			}
 
-				// 4. DB 에서 이미지 정보 삭제
-				imageRepository.deleteById(imageId);
-			});
-		}
+			// 4. DB 에서 이미지 정보 삭제
+			imageRepository.deleteById(imageId);
+		});
 
+		return imageRepository.getAllByPostId(request.postId());
+	}
+
+	@Override
+	public void deleteTempImage(Long userId) {
+		List<Image> images = imageRepository.findByPostUserUserIdAndPostStatusName(userId, "작성중");
+
+		images.forEach(image -> {
+			// 2. 실제 파일 경로 가져오기
+			Path filePath = Paths.get(uploadPath,
+				image.getUrl().replaceFirst(domainUrl, ""));
+
+			try {
+				// 3. 실제 파일 삭제
+				Files.deleteIfExists(filePath);
+			} catch (IOException e) {
+				log.error("파일 삭제 실패: {}", filePath, e);
+			}
+
+			// 4. DB 에서 이미지 정보 삭제
+			imageRepository.deleteById(image.getImageId());
+		});
 	}
 
 }
